group controller related stuff into it's own modu
