+++ src/client/src/sync/controller.rs
@@ -83,6 +83,7 @@ where
 
     #[instrument(
         skip(self, response),
+        name = "Metadata",
         fields(spec = &*S::LABEL)
     )]
     async fn dispatch_loop(mut self, mut response: AsyncResponse<WatchRequest>) {
