+++ src/spu/src/replication/follower/mod.rs
@@ -1,8 +1,7 @@
 mod state;
-mod follower_controller;
 mod api_key;
 mod peer_api;
+mod group;
 pub mod sync;
 
-pub(crate) use self::follower_controller::ReplicaFollowerController;
 pub use self::state::{FollowersState, SharedFollowersState, FollowerReplicaState};
